Completed refactoring movie_details_bloc.dart & movies_bloc.dart to use flutter_bloc
