+++ lib/ui/home/movies_event.dart
@@ -1,0 +1,3 @@
+abstract class MoviesEvent {}
+
+class GetMoviesEvent implements MoviesEvent {}
