+++ lib/ui/movie_details/movie_details_event.dart
@@ -1,0 +1,3 @@
+abstract class MovieDetailsEvent {}
+
+class GetMovieDetailsEvent implements MovieDetailsEvent {}
