+++ pubspec.yaml
@@ -40,10 +40,10 @@ dependencies:
   shared_preferences: ^2.0.5
   get: ^4.3.8
   universal_platform: 1.0.0-nullsafety
-  provider: ^5.0.0
+  provider: ^6.0.0
   hive: ^2.0.4 
   hive_flutter: ^1.0.0
-  flutter_bloc: ^7.0.0
+  flutter_bloc: ^7.3.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
